[API-92] Extract rainfall/soak/decision/util helpers from the dynamic planner.
