+++ api/schedules/dynamic/soak.ts
@@ -1,0 +1,14 @@
+/**
+ * Estimate soak time (minutes) based on infiltration rate.
+ * Lower infiltration rates require longer soak times.
+ *
+ * @param infiltrationRateMmHr - Infiltration rate in mm/hr.
+ * @returns Recommended soak time in minutes.
+ */
+export function estimateSoakMinutes(infiltrationRateMmHr: number): number {
+    if (infiltrationRateMmHr >= 20) return 15;
+    if (infiltrationRateMmHr >= 12) return 25;
+    if (infiltrationRateMmHr >= 8) return 35;
+    if (infiltrationRateMmHr >= 5) return 45;
+    return 60;
+}
