+++ api/schedules/dynamic/util.ts
@@ -1,0 +1,21 @@
+/**
+ * Clamp a value between minimum and maximum bounds.
+ *
+ * @param value - The value to clamp.
+ * @param min - Minimum allowed value.
+ * @param max - Maximum allowed value.
+ * @returns Clamped value.
+ */
+export function clampValue(value: number, min: number, max: number): number {
+    return Math.max(min, Math.min(max, value));
+}
+
+/**
+ * Round a number to one decimal place.
+ *
+ * @param value - The number to round.
+ * @returns Rounded value with one decimal place.
+ */
+export function roundTo1Decimal(value: number): number {
+    return Math.round(value * 10) / 10;
+}
